modified admin overview to several components to use both admin and agent dashboard overview page
