+++ src/components/dashboardComponents/adminOverview/OverviewHeader.jsx
@@ -2,6 +2,10 @@
 import React from 'react';
 
 const OverviewHeader = () => {
+    // ========== EXPORT HANDLER ==========
+  const exportData = (format) => {
+    alert(`Exporting as ${format.toUpperCase()}... (simulated)`);
+  };
     return (
       <div className="flex flex-col lg:flex-row justify-between items-start lg:items-center mb-8 gap-4">
         <div>
